requirements.txt
asgiref==3.8.1
Django==5.1.2
django-modeltranslation==0.19.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
pillow==11.0.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
uritemplate==4.1.1
django-debug-toolbar==4.4.6
psycopg2-binary==2.9.10
gunicorn==21.2.0
celery==5.4.0
redis==5.0.7
django-celery-beat